requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
confusable-homoglyphs==3.2.0
coverage==5.5
crispy-bootstrap5==0.6
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.5
django-allauth==0.45.0
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-countries==7.2.1
django-crispy-forms==1.13.0
django-filter==2.4.0
django-js-asset==1.2.2
django-livereload-server==0.3.2
django-registration==3.2
django-rest-auth==0.9.5
django-utils-six==2.0
django-webpack-loader==1.4.1
djangorestframework==3.12.4
future==0.18.2
httpie==2.4.0
idna==3.2
jsonparser==1.0
Markdown==3.3.4
oauthlib==3.1.1
Paperboy==1.0.1
Pillow==8.3.1
pycparser==2.20
Pygments==2.10.0
PyJWT==2.1.0
PyMySQL==0.10.1
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
pyuca==1.2
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tornado==6.1
urllib3==1.26.6
whitenoise==5.3.0
